Completed email, browserslist update
